+++ src/main/java/com/webappgroupg/SMUHiring/model/User.java
@@ -11,11 +11,11 @@ public class User {
     private String firstName;
     private String lastName;
     private String email;
-    private int phoneNumber;
+    private long phoneNumber;
     private String status;
     private Character userType;
 
-    public User(String userId, String firstName, String lastName, String email, int phoneNumber, String status) {
+    public User(String userId, String firstName, String lastName, String email, long phoneNumber, String status) {
         this.userId = userId;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -23,7 +23,7 @@ public class User {
         this.phoneNumber = phoneNumber;
         this.status = status;
     }
-    public User(String userId, String firstName, String lastName, String email, int phoneNumber) {
+    public User(String userId, String firstName, String lastName, String email, long phoneNumber) {
         this.userId = userId;
         this.firstName = firstName;
         this.lastName = lastName;
